+++ src/Controller/UserController.php
@@ -6,7 +6,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Repository\UserRepository;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
@@ -69,7 +68,7 @@ class UserController extends AbstractController
         $idCache = 'getUser';
         $jsonUser = $cache->get($idCache, function (ItemInterface $item) use ($repository, $serializer, $request, $user) {
             $item->tag("getUser");
-            $context = SerializationContext::create()->setGroups('getUser');
+            $context = SerializationContext::create()->setGroups(["getUser"]);
 
             $users = $repository->find($user);
             return $serializer->serialize($users, 'json', $context);
@@ -84,51 +83,76 @@ class UserController extends AbstractController
      */
     #[Route('/api/user/{idUser}', name: 'user.deleteUser', methods: ['DELETE'])]
     #[ParamConverter("user", class: 'App\Entity\User', options: ["id" => "idUser"])]
-    #[IsGranted('ROLE_ADMIN', message: 'Vous n\'êtes pas admin')]
     public function deleteUser(
         User $user,
+        Request $request,
+        UserRepository $userRepository,
         EntityManagerInterface $entityManager,
         TagAwareCacheInterface $cache
     ) :JsonResponse
     {
-        $cache->invalidateTags(["getUser", "getAllUsers"]);
-        $entityManager->remove($user);
-        $entityManager->flush();
-        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+        $apiToken = $request->headers->get('apiToken');
+        /** @var User $connectedUSer */
+        $connectedUSer = $userRepository->findOneBy(['apiToken' => $apiToken]);
+
+        if ($connectedUSer !== null) {
+            if ($connectedUSer->getRole() === 'admin') {
+                $cache->invalidateTags(["getUser", "getAllUsers"]);
+                $entityManager->remove($user);
+                $entityManager->flush();
+                return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+            } else {
+                return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
+            }
+        } else {
+            return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     // create
     #[Route('/api/user', name: 'user.create', methods: ['POST'])]
-    #[IsGranted('ROLE_ADMIN', message: 'Vous n\'êtes pas admin')]
     public function createUser(
         Request $request,
         SerializerInterface $serializer,
         EntityManagerInterface $entityManager,
         ValidatorInterface $validator,
+        UserRepository $userRepository,
     ) :JsonResponse
     {
-        $newUser = $serializer->deserialize(
-            $request->getContent(),
-            User::class,
-            'json');
+        $apiToken = $request->headers->get('apiToken');
+        /** @var User $connectedUSer */
+        $connectedUSer = $userRepository->findOneBy(['apiToken' => $apiToken]);
 
-        $errors = $validator->validate($newUser);
-        if ($errors->count() >0) {
-            return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+        if ($connectedUSer !== null) {
+            if ($connectedUSer->getRole() === 'admin') {
+                $newUser = $serializer->deserialize(
+                    $request->getContent(),
+                    User::class,
+                    'json');
+
+                $errors = $validator->validate($newUser);
+                if ($errors->count() >0) {
+                    return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+                }
+                $entityManager->persist($newUser);
+                $entityManager->flush();
+
+                $context = SerializationContext::create()->setGroups(["getAllUsers"]);
+
+                $jsonUser = $serializer->serialize($newUser, 'json', $context /*['groups' => 'getUser']*/);
+                return new JsonResponse($jsonUser, Response::HTTP_CREATED, [], true);
+
+            } else {
+                return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
+            }
+        } else {
+            return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
         }
-        $entityManager->persist($newUser);
-        $entityManager->flush();
-
-        $context = SerializationContext::create()->setGroups(["getAllUsers"]);
-
-        $jsonUser = $serializer->serialize($newUser, 'json', $context /*['groups' => 'getUser']*/);
-        return new JsonResponse($jsonUser, Response::HTTP_CREATED, [], true);
     }
 
     // update
     #[Route('/api/user/{idUser}', name: 'user.update', methods: ['PATCH'])]
     #[ParamConverter("user", class: 'App\Entity\User', options: ["id" => "idUser"])]
-    #[IsGranted('ROLE_ADMIN', message: 'Vous n\'êtes pas admin')]
     public function updateUser(
         User $user,
         Request $request,
@@ -136,31 +160,84 @@ class UserController extends AbstractController
         SerializerInterface $serializer,
         UserRepository $UserRepository,
         ValidatorInterface $validator,
-        UrlGeneratorInterface $urlGenerator
+        UserRepository $userRepository,
     ): JsonResponse {
-        $updateUser = $serializer->deserialize(
-            $request->getContent(),
-            User::class,
-            'json'
-        );
 
-        $user->setUsername($updateUser->getUsername() ? $updateUser->getUsername() : $user->getUsername());
-        $user->setEmail($updateUser->getEmail() ? $updateUser->getEmail() : $user->getEmail());
-        $user->setPassword($updateUser->getPassword() ? $updateUser->getPassword() : $user->getPassword());
-        $user->setPhoto($updateUser->getPhoto() ? $updateUser->getPhoto() : $user->getPhoto());
+        $apiToken = $request->headers->get('apiToken');
+        /** @var User $connectedUSer */
+        $connectedUSer = $userRepository->findOneBy(['apiToken' => $apiToken]);
 
-        $errors = $validator->validate($user);
-        if ($errors->count() >0) {
-            return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+        if ($connectedUSer !== null) {
+            if ($connectedUSer->getRole() === 'admin') {
+                $updateUser = $serializer->deserialize(
+                    $request->getContent(),
+                    User::class,
+                    'json'
+                );
+
+                $user->setUsername($updateUser->getUsername() ? $updateUser->getUsername() : $user->getUsername());
+                $user->setEmail($updateUser->getEmail() ? $updateUser->getEmail() : $user->getEmail());
+                $user->setPassword($updateUser->getPassword() ? $updateUser->getPassword() : $user->getPassword());
+                $user->setPhoto($updateUser->getPhoto() ? $updateUser->getPhoto() : $user->getPhoto());
+
+                $errors = $validator->validate($user);
+                if ($errors->count() >0) {
+                    return new JsonResponse($serializer->serialize($errors, 'json'), Response::HTTP_BAD_REQUEST, [], true);
+                }
+
+                $entityManager->persist($user);
+                $entityManager->flush();
+
+                $context = SerializationContext::create()->setGroups(["getAllUsers"]);
+
+                $jsonUSer = $serializer->serialize($user, 'json', $context);
+                return new JsonResponse($jsonUSer, Response::HTTP_CREATED, [], true);
+
+            } else {
+                return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
+            }
+        } else {
+            return new JsonResponse(['message' => 'Vous devez être connecté'], Response::HTTP_UNAUTHORIZED);
         }
+    }
 
-        $entityManager->persist($user);
+    #[Route('/api/login', name: 'user.login', methods: ['POST'])]
+    public function login(Request $request, UserRepository $userRepository, EntityManagerInterface $entityManager, SerializerInterface $serializer)
+    {
+        $context = SerializationContext::create()->setGroups(["login"]);
+
+        $data = json_decode($request->getContent(), true);
+        $username = $data['username'];
+        $password = $data['password'];
+
+        $user = $userRepository->findOneBy(['username' => $username]);
+        if ($user !== null) {
+            if ($password !== $user->getPassword()) {
+                return new JsonResponse(['message' => 'Le mot de passe est invalide'], Response::HTTP_UNAUTHORIZED);
+            }
+            $generatedToken = $serializer->serialize($user, 'json');
+
+            $user->setApiToken(base64_encode($generatedToken));
+            $entityManager->flush();
+
+            return new JsonResponse($serializer->serialize($user, 'json', $context), Response::HTTP_OK, [], true);
+        } else {
+            return new JsonResponse(['message' => 'Le nom d\'utilisateur est invalide'], Response::HTTP_UNAUTHORIZED);
+        }
+    }
+
+    #[Route('/api/logout/{idUser}', name: 'user.logout', methods: ['POST'])]
+    #[ParamConverter("user", class: 'App\Entity\User', options: ["id" => "idUser"])]
+    public function logout(
+        User $user,
+        EntityManagerInterface $entityManager,
+        TagAwareCacheInterface $cache
+    ) :JsonResponse
+    {
+        $cache->invalidateTags(["login"]);
+        $user->setApiToken(null);
         $entityManager->flush();
-
-        $context = SerializationContext::create()->setGroups(["getAllUsers"]);
-
-        $jsonBoutique = $serializer->serialize($user, 'json', $context);
-        return new JsonResponse($jsonBoutique, Response::HTTP_CREATED, [], true);
+        return new JsonResponse(['message' => 'Vous êtes déconnecté'], Response::HTTP_UNAUTHORIZED);
     }
 }
